+++ src/main/java/com/wakeUpTogetUp/togetUp/api/avatar/application/AvatarSpeechProvider.java
@@ -1,9 +1,8 @@
-package com.wakeUpTogetUp.togetUp.api.avatar;
+package com.wakeUpTogetUp.togetUp.api.avatar.application;
 
 import com.wakeUpTogetUp.togetUp.api.avatar.dto.response.AvatarSpeechResponse;
 import com.wakeUpTogetUp.togetUp.api.avatar.model.AvatarSpeech;
 import com.wakeUpTogetUp.togetUp.api.avatar.repository.AvatarSpeechRepository;
-import com.wakeUpTogetUp.togetUp.api.users.UserAvatarValidationService;
 import com.wakeUpTogetUp.togetUp.utils.mapper.EntityDtoMapper;
 import java.util.List;
 import java.util.Random;
